requirements.txt
Django==5.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
groq==1.0.0
tavily-python==0.7.21
httpx==0.28.1
pydantic==2.12.5